requirements.txt
streamlit==1.28.1
Pillow==10.0.0
